requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
Pillow==10.1.0
watchdog==3.0.0